Refactor: Move parsers back to value module
