+++ src/expression/literal.rs
@@ -9,7 +9,7 @@ use nom::{
 
 use crate::{
     basic_types::{Context, Record},
-    value::{NumericValue, Value},
+    value::{parse_float_literal, parse_integer_literal, NumericValue, Value},
 };
 
 use super::{Expression, ExpressionParseResult};
@@ -59,21 +59,6 @@ fn parse_number_literal(input: &str) -> ExpressionParseResult {
     let (i, number) = alt((parse_float_literal, parse_integer_literal))(input)?;
 
     Result::Ok((i, Box::new(Literal::Numeric(number))))
-}
-
-pub(crate) fn parse_float_literal(input: &str) -> IResult<&str, NumericValue> {
-    // Omit ? on the . to intentionally _not_ match on integers
-    let (input, matched) = re_find!(input, r"^[-+]?[0-9]*\.[0-9]+([eE][-+]?[0-9]+)?")?;
-    let number = matched.parse::<f64>().unwrap();
-
-    IResult::Ok((input, NumericValue::Float(number)))
-}
-
-pub(crate) fn parse_integer_literal(input: &str) -> IResult<&str, NumericValue> {
-    let (input, matched) = re_find!(input, r"^[-+]?[0-9]+")?;
-    let number = matched.parse::<i64>().unwrap();
-
-    IResult::Ok((input, NumericValue::Integer(number)))
 }
 
 fn parse_string_literal(input: &str) -> ExpressionParseResult {
@@ -160,52 +145,5 @@ mod tests {
             result.unwrap().1.evaluate(&context, &record),
             Value::String("hello world".to_string()),
         );
-    }
-
-    #[test]
-    fn parse_number_literals() {
-        // Integers
-        assert_eq!(
-            parse_integer_literal("123").unwrap().1,
-            NumericValue::Integer(123)
-        );
-        assert_eq!(
-            parse_integer_literal("123000").unwrap().1,
-            NumericValue::Integer(123000)
-        );
-        assert_eq!(
-            parse_integer_literal("-123").unwrap().1,
-            NumericValue::Integer(-123)
-        );
-        assert_eq!(parse_integer_literal("(123").is_err(), true);
-        // Would like this test to pass, but the distinction is implemented
-        // by the sequencing of the parsers of parse_number_literal
-        // assert_eq!(parse_integer_literal("123.45").is_err(), true);
-        assert_eq!(parse_integer_literal(".").is_err(), true);
-
-        // Floats
-        assert_eq!(
-            parse_float_literal("123.45"),
-            IResult::Ok(("", NumericValue::Float(123.45)))
-        );
-        assert_eq!(
-            parse_float_literal("123.45e-5"),
-            IResult::Ok(("", NumericValue::Float(123.45e-5)))
-        );
-        assert_eq!(
-            parse_float_literal("123.45E5"),
-            IResult::Ok(("", NumericValue::Float(123.45e5)))
-        );
-        assert_eq!(
-            parse_float_literal(".45"),
-            IResult::Ok(("", NumericValue::Float(0.45)))
-        );
-        assert_eq!(
-            parse_float_literal("-123.45"),
-            IResult::Ok(("", NumericValue::Float(-123.45)))
-        );
-        assert_eq!(parse_float_literal("a").is_err(), true);
-        assert_eq!(parse_float_literal(".").is_err(), true);
-        assert_eq!(parse_float_literal("+e").is_err(), true);
     }
 }
